fix(parser): double instead of int in replenishment or withdrawal menu, add validation also there

+++ Calculation/deposit_calculator.h
@@ -33,7 +33,7 @@ typedef struct deposit_input {
     unsigned capitalization;
     double daily_income;
     double total_profit;
-    int *account_movement;
+    double *account_movement;
 } deposit_input;
 
 typedef struct deposit_output {
